DX10 doesn't support assembly shaders - so it doesn't support the cg compiler that compiles cg to assembly, so - in order to solve this the solution is to compile cg as hlsl, the languages are almost the same so - it will work in a lot of the cases. Also  if the cg plugin registers  it starts to compile shaders and to create assembly shaders that we don't want. So  I made a little patch that unregister the cg plugin factory and register the DX10 hlsl factory instead. That is the best choice I can see for now.
I also fixed the world matrix in the render system to work.

I added lighting enabled parameter to the fixedFunc.

I added a basic texture matrix transform to the fixedFunc.

I worked more on the render to texture and got the related code to a better state.

Now a lot more of the demos work.